requirements.txt
asgiref==3.8.1
Django==5.0.7
django-appconf==1.0.6
django-browser-reload==1.13.0
django-select2==8.2.0
holidays==0.53
gunicorn==20.1.0
uvicorn==0.21.1
mysqlclient==2.2.4
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
xgboost==2.1.1
whitenoise==6.3.0
djangorestframework==3.15.2